+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Changes
 - Updated translations: `bn-IN`, `ti-ET`, `om-ET`, `bn-BD`, `te-IN`, `kn-IN`
+- Show progress when linking bp passport to patient
 
 ## On Demo
 ### Internal